Refactor payment and promo code logic; improve API structures
Replaced boolean payment flags with PaymentStatus and added gateway details. Enhanced PromoCode structure with types, usage limits, and validity periods. Refactored PaymentService, introduced transaction status updates, cancellation, and detailed exception handling.

+++ src/main/java/com/Misbra/Entity/SessionBundle.java
@@ -4,6 +4,8 @@ import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -14,6 +16,6 @@ public class SessionBundle {
     private String bundleId;
     private String bundleName;   // e.g. "Single", "10-Pack", "20-Pack"
     private Long quantity;       // how many sessions are included
-    private Double price;        // cost of the bundle
+    private BigDecimal price;        // cost of the bundle
     private boolean active;      // whether this bundle is currently for sale
 }
